requirements.txt
attrs==21.4.0
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.10.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
SQLAlchemy==1.4.28
toml==0.10.2
typing-extensions==4.0.1
webargs==8.0.1
Werkzeug==2.0.2
zipp==3.7.0
python-dateutil==2.8.2
